requirements.txt
flask
mistralai
python-dotenv
gunicorn
